renames folder and files with summary stats. removes unused code
