+++ src/components/NewLocationPage/SummaryStat/MetricValue.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Value, ValueWrapper } from './SummaryStatsBlock.style';
+import { Value, ValueWrapper } from './SummaryStat.style';
 import { CircleIcon } from '../Shared/Shared.style';
 
 const MetricValue: React.FC<{ value: string; iconColor: string }> = ({
