+++ src/components/NewLocationPage/SummaryStat/MobileSummaryStat.tsx
@@ -2,32 +2,38 @@ import React from 'react';
 import MetricSubLabel from './MetricSubLabel';
 import MetricValue from './MetricValue';
 import {
+  MetricLabel,
   StatContent,
   Row,
-  MetricLabel,
   StyledChevron,
-} from './SummaryStatsBlock.style';
+} from './SummaryStat.style';
 import { metricSubLabelText, SummaryStatProps } from './utils';
 
-const DesktopSummaryStat: React.FC<SummaryStatProps> = ({
+const MobileSummaryStat: React.FC<SummaryStatProps> = ({
   levelInfo,
   formattedValue,
   hasSubLabel,
   metricName,
+  isMobile,
   metric,
 }) => {
   return (
     <StatContent>
       <Row>
         <MetricLabel>{metricName}</MetricLabel>
-        <StyledChevron />
+        {hasSubLabel && (
+          <MetricSubLabel
+            text={metricSubLabelText[metric]}
+            isMobile={isMobile}
+          />
+        )}
       </Row>
       <Row>
         <MetricValue value={formattedValue} iconColor={levelInfo.color} />
-        {hasSubLabel && <MetricSubLabel text={metricSubLabelText[metric]} />}
+        <StyledChevron />
       </Row>
     </StatContent>
   );
 };
 
-export default DesktopSummaryStat;
+export default MobileSummaryStat;
