+++ src/components/NewLocationPage/SummaryStat/SummaryStat.tsx
