+++ src/components/NewLocationPage/SummaryStat/index.ts
